+++ edurov/core.py
@@ -10,6 +10,7 @@ from edurov.utils import warning, preexec_function
 from edurov.web import start_http_server
 
 
+
 class WebMethod(object):
     def __init__(self, video_resolution='1024x768', fps=30, server_port=8000,
                  debug=False, runtime_functions=None, index_file=None,