+++ examples_edurov/edurov_web/start.py
@@ -1,10 +1,11 @@
 import os
+import time
 
 import Pyro4
 
 from edurov import WebMethod
 from edurov.utils import detect_pi, serial_connection, send_arduino, \
-    receive_arduino
+    receive_arduino, free_drive_space
 
 if detect_pi():
     from sense_hat import SenseHat
@@ -75,13 +76,28 @@ def senser():
                           'yaw': orientation['yaw']}
 
 
+def SD_card_monitor():
+    with Pyro4.Proxy("PYRONAME:ROVSyncer") as rov:
+        while rov.run:
+            rov.sensor = {'free_space': free_drive_space()}
+            time.sleep(10)
+
+
+def recorder():
+    with Pyro4.Proxy("PYRONAME:ROVSyncer") as rov:
+        time.sleep(10)
+        rov.start_recording()
+        while rov.run:
+            pass
+
+
 def main(video_resolution='1024x768', fps=30, server_port=8000, debug=False):
     web_method = WebMethod(
         video_resolution=video_resolution,
         fps=fps,
         server_port=server_port,
         debug=debug,
-        runtime_functions=[arduino, senser],
+        runtime_functions=[arduino, senser, SD_card_monitor, recorder],
         index_file=os.path.join(os.path.dirname(__file__), 'index.html')
     )
     web_method.serve()